Moving SerdeDataProvider to its own file
