+++ components/provider_fs/Cargo.toml
@@ -32,7 +32,7 @@ extra_features = [
 ]
 
 [dependencies]
-icu_provider = { version = "0.1", path = "../provider", features = ["erased"] }
+icu_provider = { version = "0.1", path = "../provider", features = ["provider_serde"] }
 icu_locid = { version = "0.1", path = "../locid", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 erased-serde = { version = "0.3" }
