+++ components/provider_fs/src/fs_data_provider.rs
@@ -5,8 +5,8 @@ use crate::deserializer;
 use crate::error::Error;
 use crate::manifest::Manifest;
 use crate::manifest::MANIFEST_FILE;
-use icu_provider::erased::*;
 use icu_provider::prelude::*;
+use icu_provider::serde::*;
 use std::borrow::Cow;
 use std::fmt::Debug;
 use std::fs;
