+++ components/provider_fs/tests/test_file_io.rs
@@ -5,8 +5,8 @@
 use icu_locid::LanguageIdentifier;
 use icu_locid_macros::langid;
 use icu_plurals::provider::*;
-use icu_provider::erased::*;
 use icu_provider::prelude::*;
+use icu_provider::serde::*;
 use icu_provider_fs::FsDataProvider;
 use std::borrow::Cow;
